+++ src/idunno.AtProto/Repo/Models/DeleteRecordResponse.cs
@@ -3,6 +3,6 @@
 
 namespace idunno.AtProto.Repo.Models;
 
-internal record DeleteRecordResponse(Commit Commit)
+internal record DeleteRecordResponse(Commit? Commit)
 {
 }