+++ src/idunno.Bluesky/Graph/Agent/Helpers/DeleteReferenceListOptOut.cs
@@ -18,7 +18,7 @@ public partial class BlueskyAgent
     /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
     /// <exception cref="ArgumentException">Thrown when <paramref name="uri"/> is not in the correct collection.</exception>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/> is <see langword="null" />.</exception>
-    public async Task<AtProtoHttpResult<Commit>> DeleteReferenceListOptOut(AtUri uri, CancellationToken cancellationToken)
+    public async Task<AtProtoHttpResult<DeleteResult>> DeleteReferenceListOptOut(AtUri uri, CancellationToken cancellationToken)
     {
         if (!IsAuthenticated)
         {
@@ -40,7 +40,7 @@ public partial class BlueskyAgent
     /// </summary>
     /// <param name="uri">The <see cref="AtUri"/> of the reference list opt-out record.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
-    public async Task<AtProtoHttpResult<Commit>> DeleteReferenceListOptOut(AtUri uri)
+    public async Task<AtProtoHttpResult<DeleteResult>> DeleteReferenceListOptOut(AtUri uri)
     {
         return await DeleteReferenceListOptOut(uri, cancellationToken: default).ConfigureAwait(false);
     }
@@ -54,7 +54,7 @@ public partial class BlueskyAgent
     /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
     /// <exception cref="ArgumentException">Thrown when <paramref name="record"/> is not in the correct collection.</exception>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="record"/> is <see langword="null" />.</exception>
-    public async Task<AtProtoHttpResult<Commit>> DeleteReferenceListOptOut(AtProtoRepositoryRecord<ReferenceListOptOut> record, CancellationToken cancellationToken)
+    public async Task<AtProtoHttpResult<DeleteResult>> DeleteReferenceListOptOut(AtProtoRepositoryRecord<ReferenceListOptOut> record, CancellationToken cancellationToken)
     {
         if (!IsAuthenticated)
         {
@@ -76,7 +76,7 @@ public partial class BlueskyAgent
     /// </summary>
     /// <param name="record">The <see cref="AtProtoRepositoryRecord{ReferenceListOptOut}"/> of the reference list opt-out record.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
-    public async Task<AtProtoHttpResult<Commit>> DeleteReferenceListOptOut(AtProtoRepositoryRecord<ReferenceListOptOut> record)
+    public async Task<AtProtoHttpResult<DeleteResult>> DeleteReferenceListOptOut(AtProtoRepositoryRecord<ReferenceListOptOut> record)
     {
         return await DeleteReferenceListOptOut(record, cancellationToken: default).ConfigureAwait(false);
     }